save other stuff
Also saves a new idea for a logo I had. It is just

 H I
 M W

But fairly minimal. Not 100% sure on the colors, originally each square
was a different color. Can play around with it some more, maybe the
lines in the M and W should be shorter, not sure if I like them.